requirements.txt
appdirs==1.4.4
arrow==1.2.3
attrs==23.1.0
backoff==2.2.1
boto3==1.28.2
botocore==1.31.2
bravado==11.0.3
bravado-core==5.17.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
click==8.1.4
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
idna==3.4
isoduration==20.11.0
jax-jumpy==1.0.0
jmespath==1.0.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
matplotlib==3.7.0
monotonic==1.6
msgpack==1.0.5
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
pdf2image==1.16.3
Pillow==9.4.0
protobuf==4.23.4
psutil==5.9.5
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
referencing==0.29.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.8.10
s3transfer==0.6.1
seaborn==0.12.2
sentry-sdk==1.30.0
setproctitle==1.3.2
Shimmy==1.1.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
stable-baselines3==2.0.0
swagger-spec-validator==3.0.3
tensorboardX==2.6.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0
uri-template==1.3.0
urllib3==1.26.14
wandb==0.15.9
webcolors==1.13
websocket-client==1.6.1
